requirements.txt
Django==2.0.3
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.6.1
pytz==2018.3
